pyproject.toml
[project]
name = "mini-redis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "grpcio>=1.80.0",
    "grpcio-tools>=1.80.0",
]

[dependency-groups]
dev = [
    "ruff>=0.15.12",
]
